Redo the locate control Fixes #100
