+++ client/src/components/AccessCodeMap.tsx
@@ -1,5 +1,5 @@
 import { LayersControl, MapContainer, TileLayer } from "react-leaflet";
-import { LocateControl } from "./LocateControl";
+import LocateControl from "./LocateControl";
 import { GeocodeControl } from "./GeocodeControl";
 import LocationMarkers from "./LocationMarkers";
 import Control from "react-leaflet-custom-control";
@@ -14,48 +14,11 @@ export interface IAccessCodeMapProps {
 
 export default function AccessCodeMap(props: IAccessCodeMapProps) {
   const { onSignOutClick } = props;
-  const [autoLocate] = useState(() => {
-    const storedValue = sessionStorage.getItem("autoLocate");
-    if (!storedValue) {
-      return true;
-    } else {
-      return sessionStorage.getItem("autoLocate") === "true";
-    }
-  });
   const [map, setMap] = useState<L.Map | null>(null);
-
-  useEffect(() => {
-    sessionStorage.setItem("autoLocate", autoLocate.toString());
-  }, [autoLocate]);
-
-  // Getting this type definition right was a *pain*. I kept getting errors about MutableRefObject
-  // type not matching. Fix is from here: https://stackoverflow.com/a/58033283. The key is to specify
-  // both null as a valid type and null as the default. Otherwise an "undefined" sneaks in and the type
-  // won't match.
-  //
-  // Also using useEffect() for this doesn't work since the ref may not be set yet. The only
-  // way to make it work was to use the event handler method described here:
-  // https://stackoverflow.com/a/55248699. What an adventure.
-  const handleRef = (ref: L.Control.Locate | null) => {
-    if (autoLocate && ref) {
-      ref.start();
-    }
-  };
 
   const storeMapRef = (ref: L.Map | null) => {
     if (ref) {
       setMap(ref);
-
-      // Add listeners for the locateactivate and locatedeactivate events so we can
-      // persist the setting in storage and restore it when the map is reloaded.
-      // I'd normally use useMapEvents() for this but it isn't accessible at any
-      // level other than *inside* the MapContainer.
-      map?.addEventListener("locateactivate", (e) => {
-        sessionStorage.setItem("autoLocate", "true");
-      });
-      map?.addEventListener("locatedeactivate", (e) => {
-        sessionStorage.setItem("autoLocate", "false");
-      });
     }
   };
 
@@ -92,15 +55,13 @@ export default function AccessCodeMap(props: IAccessCodeMapProps) {
       <GeocodeControl />
       <LocateControl
         position="topright"
-        ref={handleRef}
-        options={{
-          setView: "always",
-          // Issue 101: High accuracy causes the dot to jump around too much when using on a mobile
-          // device so disable it. Leaving the code here with it as false so in the future I remember
-          // how to turn it on if I want to.
-          locateOptions: {
-            enableHighAccuracy: false,
-          },
+        setView="always"
+        autoStart
+        // Issue 101: High accuracy causes the dot to jump around too much when using on a mobile
+        // device so disable it. Leaving the code here with it as false so in the future I remember
+        // how to turn it on if I want to.
+        locateOptions={{
+          enableHighAccuracy: false,
         }}
       />
       <LocationMarkers />
