+++ client/src/components/LocateControl.tsx
@@ -1,24 +1,49 @@
-import { createControlComponent } from "@react-leaflet/core";
 import * as L from "leaflet";
 import "leaflet.locatecontrol";
 import "leaflet.locatecontrol/dist/L.Control.Locate.css";
+import { EffectCallback, useEffect, useState } from "react";
+import { useMap } from "react-leaflet";
 
-// This code comes directly from https://stackoverflow.com/a/75567918
-interface P extends L.ControlOptions {
-  position?: L.ControlPosition;
-  options?: L.Control.LocateOptions;
+interface ILocateControlProps extends L.Control.LocateOptions {
+  autoStart?: boolean;
 }
 
-const { Locate } = L.Control;
+const LocateControl = (props: ILocateControlProps) => {
+  const { autoStart, ...rest } = props;
+  const [savedAutoLocateState] = useState(() => {
+    const storedValue = sessionStorage.getItem("autoLocate");
+    if (!storedValue) {
+      return true;
+    } else {
+      return sessionStorage.getItem("autoLocate") === "true";
+    }
+  });
+  const map = useMap();
 
-function createLocateInstance(props: P) {
-  const instance = new Locate({
-    position: props.position,
-    ...props.options,
-    //    ...props,
+  // Add listeners for the locateactivate and locatedeactivate events so we can
+  // persist the setting in storage and restore it when the map is reloaded.
+  // I'd normally use useMapEvents() for this but it isn't accessible at any
+  // level other than *inside* the MapContainer.
+  map?.addEventListener("locateactivate", (e) => {
+    sessionStorage.setItem("autoLocate", "true");
+  });
+  map?.addEventListener("locatedeactivate", (e) => {
+    sessionStorage.setItem("autoLocate", "false");
   });
 
-  return instance;
-}
+  useEffect((): ReturnType<EffectCallback> => {
+    const instance = new L.Control.Locate({
+      ...rest,
+    });
 
-export const LocateControl = createControlComponent(createLocateInstance);
+    map.addControl(instance);
+
+    return () => {
+      map.removeControl(instance);
+    };
+  }, [map, rest]);
+
+  return null;
+};
+
+export default LocateControl;
